That doesn't need to be so wide

+++ fbink_device_id.c
@@ -191,7 +191,7 @@ static uint32_t
 
 	// Hi, my name is Neo. I know pointers! (Or not.)
 	for (const char* p = num; *p != '\0'; p++) {
-		for (size_t i = 0; tbl[i] != '\0'; i++) {
+		for (uint8_t i = 0; tbl[i] != '\0'; i++) {
 			if (*p == tbl[i]) {
 				result = result * base + i;
 				break;
